+++ src/main/java/walkingkooka/spreadsheet/SpreadsheetContextSupplierTesting2.java
@@ -1,0 +1,38 @@
+/*
+ * Copyright 2019 Miroslav Pokorny (github.com/mP1)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package walkingkooka.spreadsheet;
+
+import org.junit.jupiter.api.Test;
+import walkingkooka.reflect.ClassTesting;
+
+import static org.junit.jupiter.api.Assertions.assertThrows;
+
+public interface SpreadsheetContextSupplierTesting2<C extends SpreadsheetContextSupplier> extends SpreadsheetContextSupplierTesting,
+    ClassTesting<C> {
+
+    @Test
+    default void testSpreadsheetContextWithNullFails() {
+        assertThrows(
+            NullPointerException.class,
+            () -> this.createSpreadsheetContextSupplier()
+                .spreadsheetContext(null)
+        );
+    }
+
+    C createSpreadsheetContextSupplier();
+}
